refactor: share one FormPage and FormModal across the create surfaces
