+++ src/app/components/page/FormPage.tsx
@@ -1,0 +1,124 @@
+import type { ReactNode } from 'react';
+import { useEffect, useState } from 'react';
+import { useSetAtom } from 'jotai';
+import { Box, IconButton, Scroll, Text, color, config, toRem } from 'folds';
+import { SquaresFour, composerIcon, sizedIcon, X } from '$components/icons/phosphor';
+import { ScreenSize, useScreenSizeContext } from '$hooks/useScreenSize';
+import { useSetting } from '$state/hooks/settings';
+import { settingsAtom } from '$state/settings';
+import { isResizingSidebarAtom } from '$state/isResizingSidebar';
+import { SidebarResizer } from '$pages/client/sidebar/SidebarResizer';
+import { UserQuickTools } from '$pages/client/sidebar/UserQuickTools';
+import {
+  Page,
+  PageContent,
+  PageContentCenter,
+  PageHero,
+  PageHeroSection,
+  PageNav,
+  PageNavHeader,
+} from './Page';
+
+type FormPageProps = {
+  title: string;
+  subTitle: string;
+  closeLabel: string;
+  onClose: () => void;
+  children: ReactNode;
+};
+
+export function FormPage({ title, subTitle, closeLabel, onClose, children }: FormPageProps) {
+  const setIsResizingSidebar = useSetAtom(isResizingSidebarAtom);
+  const [roomSidebarWidth, setRoomSidebarWidth] = useSetting(settingsAtom, 'roomSidebarWidth');
+  const [curWidth, setCurWidth] = useState(roomSidebarWidth);
+
+  useEffect(() => {
+    setCurWidth(roomSidebarWidth);
+  }, [roomSidebarWidth]);
+
+  const screenSize = useScreenSizeContext();
+  const isMobile = screenSize === ScreenSize.Mobile;
+  const hideText = curWidth <= 80 && !isMobile;
+  const [oldSidebar] = useSetting(settingsAtom, 'oldSidebar');
+
+  return (
+    <>
+      {!isMobile && (
+        <Box
+          shrink="No"
+          style={{
+            position: 'relative',
+            width: toRem(curWidth),
+            borderRight: 'solid',
+            borderColor: color.SurfaceVariant.ContainerLine,
+            borderWidth: `0 ${config.borderWidth.B300} 0 0`,
+          }}
+        >
+          <PageNav>
+            <PageNavHeader size="600">
+              <Box grow="Yes" gap="300" justifyContent="Center">
+                {hideText ? (
+                  sizedIcon(SquaresFour, '200', { filled: true })
+                ) : (
+                  <Box grow="Yes">
+                    <Text size="H4" truncate align="Center">
+                      {title}
+                    </Text>
+                  </Box>
+                )}
+              </Box>
+            </PageNavHeader>
+            <SidebarResizer
+              setCurWidth={setCurWidth}
+              sidebarWidth={roomSidebarWidth}
+              setSidebarWidth={setRoomSidebarWidth}
+              instep={50}
+              outstep={190}
+              minValue={50}
+              maxValue={500}
+              setAnnouncement={setIsResizingSidebar}
+            />
+          </PageNav>
+          {!oldSidebar && <UserQuickTools width={curWidth + 66} compact={false} />}
+        </Box>
+      )}
+      <Page>
+        <Box grow="Yes" direction="Column" style={{ background: color.Background.Container }}>
+          {isMobile && (
+            <PageNav>
+              <PageNavHeader size="600">
+                <Box grow="Yes" gap="300" alignItems="Center">
+                  <Box grow="Yes" style={{ flexBasis: 0 }} />
+                  <Text size="H4" align="Center" truncate>
+                    {title}
+                  </Text>
+                  <Box grow="Yes" style={{ flexBasis: 0 }} justifyContent="End">
+                    <IconButton size="300" radii="300" aria-label={closeLabel} onClick={onClose}>
+                      {composerIcon(X)}
+                    </IconButton>
+                  </Box>
+                </Box>
+              </PageNavHeader>
+            </PageNav>
+          )}
+          <Scroll hideTrack visibility="Hover">
+            <PageContent>
+              <PageContentCenter>
+                <PageHeroSection>
+                  <Box direction="Column" gap="700">
+                    <PageHero
+                      icon={sizedIcon(SquaresFour, '600')}
+                      title={title}
+                      subTitle={subTitle}
+                    />
+                    {children}
+                  </Box>
+                </PageHeroSection>
+              </PageContentCenter>
+            </PageContent>
+          </Scroll>
+        </Box>
+      </Page>
+    </>
+  );
+}
